4.21*: blogilistan laajennus, step9

+++ controllers/blogs.js
@@ -42,8 +42,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const user = request.token === undefined
+    ? false
+    : jwt.verify(request.token, process.env.SECRET)
+  const blog = await Blog.findById(request.params.id)
+  if( user.id && blog.user && blog.user.toString() === user.id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    return response.status(204).end()
+  }
+  return response.status(401).json({ error: 'unauthorized action: missing token or unauthorized user' })
 })
 
 blogsRouter.put('/:id', async (request, response) => {
